requirements.txt
flask>=3.0.0
anthropic>=0.40.0
pillow>=10.0.0
pillow-heif>=0.15.0
openpyxl>=3.1.0
requests>=2.31.0
